backend/requirements.txt
fastapi
uvicorn
websockets
requests
aiofiles
python-multipart